python/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.0
azure-storage-blob==12.13.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
frozenlist==1.3.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.1
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
